+++ src/icemac/ab/calendar/browser/renderer/table.py
@@ -2,7 +2,7 @@
 from __future__ import unicode_literals
 
 from .base import Calendar
-from .interfaces import UnknownLanguageError
+from ..interfaces import UnknownLanguageError
 from icemac.addressbook.browser.base import can_access_uri_part
 from icemac.addressbook.i18n import _
 import datetime