+++ src/icemac/ab/calendar/browser/tests/test_calendar.py
@@ -4,7 +4,7 @@ from datetime import timedelta, date
 from icemac.ab.calendar.browser.calendar import EventDescription, hyphenated
 from icemac.ab.calendar.browser.calendar import TabularCalendar
 from icemac.ab.calendar.browser.interfaces import IEventDescription
-from icemac.ab.calendar.browser.renderer.interfaces import UnknownLanguageError
+from icemac.ab.calendar.browser.interfaces import UnknownLanguageError
 from icemac.ab.calendar.interfaces import ICalendarDisplaySettings, ICalendar
 from icemac.ab.calendar.interfaces import IEvent, IRecurringEvent
 from icemac.ab.calendar.testing import get_recurred_event