+++ CHANGES.rst
@@ -12,7 +12,8 @@ Backward incompatible changes
   The previous functionality of ``get_events()`` is now provided by
   ``get_events_for_month()``
 
-- Move ``.browser.renderer.interfaces.IEventDescription`` to
+- Move ``.browser.renderer.interfaces.IEventDescription`` and
+  ``.browser.renderer.interfaces.UnknownLanguageError`` to
   ``.browser.interfaces``.